# functions/docs/openapi.yaml
openapi: 3.0.0
info:
  title: CollabCanvas API
  version: 1.0.0
  description: |
    REST API for CollabCanvas - A collaborative canvas platform for real-time design and AI integration.
    
    ## Authentication
    All endpoints (except `/api/health`) require authentication using Bearer tokens.
    Include your API token in the Authorization header:
    ```
    Authorization: Bearer YOUR_API_TOKEN
    ```
    
    ## Rate Limits
    - Read operations: 200 requests per minute
    - Write operations: 50 requests per minute
    
    ## Error Responses
    All errors follow a consistent format:
    ```json
    {
      "error": {
        "message": "Error description",
        "code": "ERROR_CODE"
      }
    }
    ```
  contact:
    name: CollabCanvas Support
    email: support@collabcanvas.example.com
servers:
  - url: https://your-project.web.app/api
    description: Production server
  - url: http://localhost:5001/your-project/us-central1/api
    description: Local development server
...
paths:
...
  /tokens/generate:
    post:
      summary: Generate API token
      description: Generate a new API token (requires Firebase Auth)
      tags: [Tokens]
      security:
        - FirebaseAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - canvasId
              properties:
                name:
                  type: string
                  example: AI Assistant Token
                canvasId:
                  type: string
                permissions:
                  type: array
                  items:
                    type: string
                    enum: [read, create_objects, update_objects, delete_objects]
                  default: [read, create_objects, update_objects, delete_objects]
                expiresInDays:
                  type: integer
                  default: 90
      responses:
        '201':
          description: Token generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: The plaintext token (shown only once)
                  tokenId:
                    type: string
                  name:
                    type: string
                  canvasId:
                    type: string
                  canvasName:
                    type: string
                  permissions:
                    type: array
                    items:
                      type: string
                  expiresAt:
                    type: string
                    format: date-time
                  warning:
                    type: string
  /tokens/{tokenId}:
    delete:
      summary: Revoke token
      description: Revoke an API token (requires Firebase Auth)
      tags: [Tokens]
      security:
        - FirebaseAuth: []
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Token revoked
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Token
      description: API token for external/AI access
    FirebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token for user authentication
...